:ok_hand: [#3279] Polish PR to make it ready for merge
