+++ src/openforms/registrations/contrib/zgw_apis/tests/test_admin.py
@@ -4,7 +4,6 @@ from django.utils.translation import gettext as _
 import requests
 import requests_mock
 from django_webtest import WebTest
-from privates.test import temp_private_root
 from zgw_consumers.api_models.constants import VertrouwelijkheidsAanduidingen
 
 from openforms.accounts.tests.factories import SuperUserFactory
@@ -14,11 +13,12 @@ from .factories import ZGWApiGroupConfigFactory
 
 
 @disable_2fa
-@temp_private_root()
 class ZGWApiGroupConfigAdminTests(WebTest):
-    csrf_checks = False
-
-    def test_admin_while_services_are_down(self):
+    @requests_mock.Mocker()
+    def test_admin_while_services_are_down(self, m):
+        m.register_uri(
+            requests_mock.ANY, requests_mock.ANY, exc=requests.ConnectTimeout
+        )
         zgw_group = ZGWApiGroupConfigFactory.create(
             zrc_service__api_root="https://zaken-1.nl/api/v1/",
             zrc_service__oas="https://zaken-1.nl/api/v1/schema/openapi.yaml",
@@ -33,34 +33,31 @@ class ZGWApiGroupConfigAdminTests(WebTest):
         )
         superuser = SuperUserFactory.create()
 
-        with requests_mock.Mocker() as m:
-            m.register_uri(
-                requests_mock.ANY,
-                requests_mock.ANY,
-                exc=requests.exceptions.ConnectTimeout,
-            )
-            response = self.app.get(
-                reverse("admin:zgw_apis_zgwapigroupconfig_change", args=(1,)),
-                user=superuser,
-            )
-            self.assertEqual(response.status_code, 200)
+        response = self.app.get(
+            reverse("admin:zgw_apis_zgwapigroupconfig_change", args=(zgw_group.pk,)),
+            user=superuser,
+        )
 
-            form = response.forms["zgwapigroupconfig_form"]
+        self.assertEqual(response.status_code, 200)
+        form = response.forms["zgwapigroupconfig_form"]
+        self.assertEqual(
+            form["zaaktype"].value, "https://catalogi-1.nl/api/v1/zaaktypen/1"
+        )
+        error_node = response.pyquery(
+            ".field-zaaktype .openforms-error-widget .openforms-error-widget__column small"
+        )
+        self.assertEqual(
+            error_node.text(),
+            _(
+                "Could not load data - enable and check the request "
+                "logs for more details."
+            ),
+        )
 
-            self.assertEqual(form["zaaktype"].value, zgw_group.zaaktype)
-
-            error_node = response.pyquery(
-                ".field-zaaktype .openforms-error-widget .openforms-error-widget__column small"
-            )
-            self.assertEqual(
-                error_node.text(),
-                _(
-                    "Could not load data - enable and check the request logs for more details."
-                ),
-            )
-
+        with self.subTest("submitting form doesn't clear configuration"):
             form.submit()
 
+            zgw_group.refresh_from_db()
             self.assertEqual(
-                form["zaaktype"].value, "https://catalogi-1.nl/api/v1/zaaktypen/1"
+                zgw_group.zaaktype, "https://catalogi-1.nl/api/v1/zaaktypen/1"
             )
